requirements.txt
ccxt>=4.3.0
pandas>=2.0.0
pandas-ta>=0.3.14b
python-dotenv>=1.0.0
requests>=2.31.0
google-genai>=1.0.0
